Week 3 Mid-Term Pull Request

+++ ssjd/src/main/java/com/daou/ssjd/domain/repository/ProblemsRepository.java
@@ -3,6 +3,6 @@ package com.daou.ssjd.domain.repository;
 import com.daou.ssjd.domain.entity.Problems;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProblemsRepository extends JpaRepository<Problems, Long> {
-    Problems findByProblemId(long problemId);
+public interface ProblemsRepository extends JpaRepository<Problems, Integer> {
+    Problems findByProblemId(int problemId);
 }
